requirements.txt
Flask_Cors==3.0.10
Flask==1.1.2
Flask_SQLAlchemy==2.4.3
Werkzeug==1.0.1
PyJWT==2.1.0
python-decouple==3.4
